+++ babel.config.js
@@ -7,6 +7,7 @@ module.exports = function(api) {
     ],
     plugins: [
       ['@babel/plugin-transform-flow-strip-types'],
+      'react-native-reanimated/plugin',
     ],
   };
 };