+++ react-native.config.js
@@ -3,6 +3,7 @@ module.exports = {
     ios: {},
     android: {
       sourceDir: './android',
+      packageName: 'com.anonymous.soundtherapyapp',
     },
   },
   assets: ['./node_modules/react-native-vector-icons/Fonts'],